+++ test/global.test.ts
@@ -5,9 +5,13 @@
 
 import { TestOutputChannel, TestUserInput } from '@microsoft/vscode-azext-dev';
 import * as vscode from 'vscode';
-import { ext, registerOnActionStartHandler } from '../extension.bundle';
+import { ext, registerOnActionStartHandler, settingUtils } from '../extension.bundle';
 
-export let longRunningTestsEnabled: boolean;
+const longRunningLocalTestsEnabled: boolean = !/^(false|0)?$/i.test(process.env.AzCode_EnableLongRunningTestsLocal || '');
+const longRunningRemoteTestsEnabled: boolean = !/^(false|0)?$/i.test(process.env.AzCode_UseAzureFederatedCredentials || '');
+
+export const longRunningTestsEnabled: boolean = longRunningLocalTestsEnabled || longRunningRemoteTestsEnabled;
+export const userSettings: { key: string, value: unknown }[] = [];
 
 // Runs before all tests
 suiteSetup(async function (this: Mocha.Context): Promise<void> {
@@ -22,6 +26,17 @@ suiteSetup(async function (this: Mocha.Context): Promise<void> {
         // Use `TestUserInput` by default so we get an error if an unexpected call to `context.ui` occurs, rather than timing out
         context.ui = new TestUserInput(vscode);
     });
+    const groupBySetting = settingUtils.getWorkspaceSetting('groupBy')
+    userSettings.push({ key: 'groupBy', value: groupBySetting });
 
-    longRunningTestsEnabled = !/^(false|0)?$/i.test(process.env.ENABLE_LONG_RUNNING_TESTS || '');
+    const deleteConfirmationSetting = settingUtils.getWorkspaceSetting('deleteConfirmation');
+    userSettings.push({ key: 'deleteConfirmation', value: deleteConfirmationSetting });
+});
+
+suiteTeardown(async function (this: Mocha.Context): Promise<void> {
+    for (const setting of userSettings) {
+        // reset the settings to their original values
+        console.debug(`Resetting setting '${setting.key}' to '${setting.value}'`);
+        await settingUtils.updateGlobalSetting(setting.key, setting.value);
+    }
 });